+++ api/components/elements.js
@@ -3043,8 +3043,26 @@ module.exports = {
         required: false,
         private: false,
       },
+      clearOnRefetch: {
+        required: 'false',
+        default: 'true',
+        localized: false,
+        types: [
+          'boolean',
+        ],
+        required: false,
+        private: false,
+      },
     },
     data: {
+      watchers: {
+        types: [
+          'array',
+        ],
+        description: 'Stores watchers for fields from which dynamic values for endpoints are retrieved.',
+        default: '[]',
+        private: true,
+      },
       active: {
         types: [
           'boolean',
@@ -3516,6 +3534,10 @@ module.exports = {
       },
     },
     methods: {
+      resolveOptions: {
+        description: 'Resolves items.',
+        private: true,
+      },
       updateItems: {
         description: 'Fetches & updates items when using `async` items.',
         returns: 'void',
@@ -3529,6 +3551,39 @@ module.exports = {
           },
         },
         private: false,
+      },
+      cleanupValue: {
+        description: 'Removes any value that is not among the newly fetches option list after async resolve.',
+        params: {
+          values: {
+            types: [
+              'array',
+            ],
+            required: 'true',
+            description: 'the list of option values',
+          },
+        },
+        private: true,
+      },
+      resolveUrlAndSetWatchers: {
+        description: 'Resolves the endpoint url with field values and sets watchers for those fields.',
+        params: {
+          url: {
+            types: [
+              'string',
+            ],
+            required: 'true',
+            description: 'the base url potentially containing variable names',
+          },
+          updateItems: {
+            types: [
+              'function',
+            ],
+            required: 'true',
+            description: 'the method that triggers item updates',
+          },
+        },
+        private: true,
       },
       activate: {
         description: 'Sets the `active` property of the element to `true`.',
@@ -18755,6 +18810,16 @@ module.exports = {
         required: false,
         private: false,
       },
+      clearOnRefetch: {
+        required: 'false',
+        default: 'true',
+        localized: false,
+        types: [
+          'boolean',
+        ],
+        required: false,
+        private: false,
+      },
       delay: {
         required: 'false',
         default: '-1',
@@ -18901,6 +18966,14 @@ module.exports = {
       },
     },
     data: {
+      watchers: {
+        types: [
+          'array',
+        ],
+        description: 'Stores watchers for fields from which dynamic values for endpoints are retrieved.',
+        default: '[]',
+        private: true,
+      },
       active: {
         types: [
           'boolean',
@@ -19415,6 +19488,10 @@ module.exports = {
       },
     },
     methods: {
+      resolveOptions: {
+        description: 'Resolves items.',
+        private: true,
+      },
       updateItems: {
         description: 'Fetches & updates select options when using `async` options. Receives [`el$`](#property-el) as first param.',
         returns: 'void',
@@ -19428,6 +19505,39 @@ module.exports = {
           },
         },
         private: false,
+      },
+      cleanupValue: {
+        description: 'Removes any value that is not among the newly fetches option list after async resolve.',
+        params: {
+          values: {
+            types: [
+              'array',
+            ],
+            required: 'true',
+            description: 'the list of option values',
+          },
+        },
+        private: true,
+      },
+      resolveUrlAndSetWatchers: {
+        description: 'Resolves the endpoint url with field values and sets watchers for those fields.',
+        params: {
+          url: {
+            types: [
+              'string',
+            ],
+            required: 'true',
+            description: 'the base url potentially containing variable names',
+          },
+          updateItems: {
+            types: [
+              'function',
+            ],
+            required: 'true',
+            description: 'the method that triggers item updates',
+          },
+        },
+        private: true,
       },
       activate: {
         description: 'Sets the `active` property of the element to `true`.',
@@ -23456,8 +23566,26 @@ module.exports = {
         required: false,
         private: false,
       },
+      clearOnRefetch: {
+        required: 'false',
+        default: 'true',
+        localized: false,
+        types: [
+          'boolean',
+        ],
+        required: false,
+        private: false,
+      },
     },
     data: {
+      watchers: {
+        types: [
+          'array',
+        ],
+        description: 'Stores watchers for fields from which dynamic values for endpoints are retrieved.',
+        default: '[]',
+        private: true,
+      },
       active: {
         types: [
           'boolean',
@@ -23929,6 +24057,10 @@ module.exports = {
       },
     },
     methods: {
+      resolveOptions: {
+        description: 'Resolves items.',
+        private: true,
+      },
       updateItems: {
         description: 'Fetches & updates items when using `async` items.',
         returns: 'void',
@@ -23942,6 +24074,39 @@ module.exports = {
           },
         },
         private: false,
+      },
+      cleanupValue: {
+        description: 'Removes any value that is not among the newly fetches option list after async resolve.',
+        params: {
+          values: {
+            types: [
+              'array',
+            ],
+            required: 'true',
+            description: 'the list of option values',
+          },
+        },
+        private: true,
+      },
+      resolveUrlAndSetWatchers: {
+        description: 'Resolves the endpoint url with field values and sets watchers for those fields.',
+        params: {
+          url: {
+            types: [
+              'string',
+            ],
+            required: 'true',
+            description: 'the base url potentially containing variable names',
+          },
+          updateItems: {
+            types: [
+              'function',
+            ],
+            required: 'true',
+            description: 'the method that triggers item updates',
+          },
+        },
+        private: true,
       },
       activate: {
         description: 'Sets the `active` property of the element to `true`.',
@@ -25257,6 +25422,16 @@ module.exports = {
         required: false,
         private: false,
       },
+      clearOnRefetch: {
+        required: 'false',
+        default: 'true',
+        localized: false,
+        types: [
+          'boolean',
+        ],
+        required: false,
+        private: false,
+      },
       delay: {
         required: 'false',
         default: '-1',
@@ -25403,6 +25578,14 @@ module.exports = {
       },
     },
     data: {
+      watchers: {
+        types: [
+          'array',
+        ],
+        description: 'Stores watchers for fields from which dynamic values for endpoints are retrieved.',
+        default: '[]',
+        private: true,
+      },
       active: {
         types: [
           'boolean',
@@ -25917,6 +26100,10 @@ module.exports = {
       },
     },
     methods: {
+      resolveOptions: {
+        description: 'Resolves items.',
+        private: true,
+      },
       updateItems: {
         description: 'Fetches & updates select options when using `async` options. Receives [`el$`](#property-el) as first param.',
         returns: 'void',
@@ -25930,6 +26117,39 @@ module.exports = {
           },
         },
         private: false,
+      },
+      cleanupValue: {
+        description: 'Removes any value that is not among the newly fetches option list after async resolve.',
+        params: {
+          values: {
+            types: [
+              'array',
+            ],
+            required: 'true',
+            description: 'the list of option values',
+          },
+        },
+        private: true,
+      },
+      resolveUrlAndSetWatchers: {
+        description: 'Resolves the endpoint url with field values and sets watchers for those fields.',
+        params: {
+          url: {
+            types: [
+              'string',
+            ],
+            required: 'true',
+            description: 'the base url potentially containing variable names',
+          },
+          updateItems: {
+            types: [
+              'function',
+            ],
+            required: 'true',
+            description: 'the method that triggers item updates',
+          },
+        },
+        private: true,
       },
       activate: {
         description: 'Sets the `active` property of the element to `true`.',
@@ -34909,6 +35129,16 @@ module.exports = {
         required: false,
         private: false,
       },
+      clearOnRefetch: {
+        required: 'false',
+        default: 'true',
+        localized: false,
+        types: [
+          'boolean',
+        ],
+        required: false,
+        private: false,
+      },
       delay: {
         required: 'false',
         default: '-1',
@@ -35065,6 +35295,14 @@ module.exports = {
       },
     },
     data: {
+      watchers: {
+        types: [
+          'array',
+        ],
+        description: 'Stores watchers for fields from which dynamic values for endpoints are retrieved.',
+        default: '[]',
+        private: true,
+      },
       active: {
         types: [
           'boolean',
@@ -35579,6 +35817,10 @@ module.exports = {
       },
     },
     methods: {
+      resolveOptions: {
+        description: 'Resolves items.',
+        private: true,
+      },
       updateItems: {
         description: 'Fetches & updates select options when using `async` options. Receives [`el$`](#property-el) as first param.',
         returns: 'void',
@@ -35592,6 +35834,39 @@ module.exports = {
           },
         },
         private: false,
+      },
+      cleanupValue: {
+        description: 'Removes any value that is not among the newly fetches option list after async resolve.',
+        params: {
+          values: {
+            types: [
+              'array',
+            ],
+            required: 'true',
+            description: 'the list of option values',
+          },
+        },
+        private: true,
+      },
+      resolveUrlAndSetWatchers: {
+        description: 'Resolves the endpoint url with field values and sets watchers for those fields.',
+        params: {
+          url: {
+            types: [
+              'string',
+            ],
+            required: 'true',
+            description: 'the base url potentially containing variable names',
+          },
+          updateItems: {
+            types: [
+              'function',
+            ],
+            required: 'true',
+            description: 'the method that triggers item updates',
+          },
+        },
+        private: true,
       },
       activate: {
         description: 'Sets the `active` property of the element to `true`.',