+++ src/components/elements/RadiogroupElement.js
@@ -69,6 +69,11 @@ export default {
       type: [Array],
       default: () => ([])
     },
+    clearOnRefetch: {
+      type: [Boolean],
+      required: false,
+      default: true,
+    },
   },
   setup(props, context) {
     context.features = [
@@ -81,11 +86,11 @@ export default {
       useEvents,
       useBaseElement,
       useDisabled,
-      useAsyncItems,
       useDefault,
+      useValue,
+      useAsyncItems,
       useConditions,
       useValidation,
-      useValue,
       useData,
       useLabel,
       useGenericName,