+++ src/components/elements/TagsElement.js
@@ -6,7 +6,6 @@ import useLayout from './../../composables/elements/useLayout'
 import useInput from './../../composables/elements/useInput'
 import usePath from './../../composables/elements/usePath'
 import useConditions from './../../composables/useConditions'
-import useData from './../../composables/elements/useData'
 import useDefault from './../../composables/elements/useDefault'
 import useValidation from './../../composables/elements/useValidation'
 import useLabel from './../../composables/elements/useLabel'
@@ -25,7 +24,6 @@ import useValue from './../../composables/elements/useValue'
 import useWatchValue from './../../composables/elements/useWatchValue'
 import useFloating from './../../composables/elements/useFloating'
 import useLoading from './../../composables/elements/useLoading'
-import useAsyncItems from './../../composables/elements/useAsyncItems'
 import useA11y from './../../composables/elements/useA11y'
 import useFocus from './../../composables/elements/useFocus'
 import usePlaceholder from './../../composables/elements/usePlaceholder'
@@ -35,6 +33,8 @@ import { array as useEmpty } from './../../composables/elements/useEmpty'
 import { tags as useOptions } from './../../composables/elements/useOptions'
 import { tags as useBaseElement } from './../../composables/elements/useBaseElement'
 import { tags as useFocused } from './../../composables/elements/useFocused'
+import { tags as useData } from './../../composables/elements/useData'
+import { tags as useAsyncItems } from './../../composables/elements/useAsyncItems'
 
 import BaseElement from './../../mixins/BaseElement'
 import HasView from './../../mixins/HasView'
@@ -277,6 +277,11 @@ export default {
       default: false,
       native: false,
     },
+    clearOnRefetch: {
+      type: [Boolean],
+      required: false,
+      default: true,
+    },
 
     delay: {
       type: [Number],
@@ -383,8 +388,8 @@ export default {
       useValidation,
       useLoading,
       useOptions,
-      useAsyncItems,
       useValue,
+      useAsyncItems,
       useConditions,
       useData,
       useEmpty,